+++ app/preload/filebrowser.js
@@ -22,8 +22,8 @@ const {joinPath, basePath} = require("../util")
 
 const styling = `body {color: var(--fg, #eee);display: flex;
     font: 14px monospace;line-height: 1.5;margin: 0;}
-main {margin: 3em auto;width: 50vw;background: #7772;padding: 3em;
-    min-width: 300px;overflow: hidden;text-overflow: ellipsis;}
+main {margin: 3em auto;width: 50vw;background: #7772;min-width: 300px;
+    padding: 3em;overflow: visible;text-overflow: ellipsis;height: fit-content;}
 h2 {font-size: 2em;margin: 0 0 1em;}
 .dir, .file {margin: .7em;cursor: pointer;}
 .dir {font-weight: bold;color: var(--suggestions-file, #ffb);}